refactor price handling and move admin file upload endpoint

+++ src/main/java/org/plishka/backend/dto/admin/product/AdminProductDetailDto.java
@@ -1,6 +1,5 @@
 package org.plishka.backend.dto.admin.product;
 
-import java.math.BigDecimal;
 import java.util.List;
 import org.plishka.backend.dto.product.CategoryDto;
 import org.plishka.backend.dto.product.ProductMediaDto;
@@ -9,7 +8,7 @@ public record AdminProductDetailDto(
         Long productId,
         String name,
         String description,
-        BigDecimal price,
+        Long price,
         CategoryDto category,
         List<ProductMediaDto> media
 ) {
