+++ src/main/java/org/plishka/backend/dto/cart/CartSummaryDto.java
@@ -1,10 +1,9 @@
 package org.plishka.backend.dto.cart;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record CartSummaryDto(
         List<CartItemSummaryDto> items,
-        BigDecimal totalPrice
+        Long totalPrice
 ) {
 }
