+++ src/main/java/org/plishka/backend/dto/home/HomePageProductDto.java
@@ -1,6 +1,5 @@
 package org.plishka.backend.dto.home;
 
-import java.math.BigDecimal;
 import org.plishka.backend.dto.product.CategoryDto;
 import org.plishka.backend.dto.product.ProductMediaPreviewDto;
 
@@ -8,7 +7,7 @@ public record HomePageProductDto(
         Long productId,
         String name,
         CategoryDto category,
-        BigDecimal price,
+        Long price,
         ProductMediaPreviewDto primaryMedia
 ) {
 }
