+++ src/main/java/org/plishka/backend/dto/order/OrderSummaryDto.java
@@ -1,12 +1,11 @@
 package org.plishka.backend.dto.order;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 public record OrderSummaryDto(
         Long orderId,
         String orderNumber,
-        BigDecimal totalPrice,
+        Long totalPrice,
         Instant createdAt
 ) {
 }
