+++ src/main/java/org/plishka/backend/dto/product/ProductDetailDto.java
@@ -1,13 +1,12 @@
 package org.plishka.backend.dto.product;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record ProductDetailDto(
         Long productId,
         String name,
         String description,
-        BigDecimal price,
+        Long price,
         CategoryDto category,
         List<ProductMediaDto> media
 ) {
